completed country, province, district, local CRUD with all API testing

+++ src/main/java/com/example/cv_generator/repository/LocalLevelRepository.java
@@ -3,5 +3,5 @@ package com.example.cv_generator.repository;
 import com.example.cv_generator.entity.LocalLevel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LocalLevelRepository extends JpaRepository<LocalLevel,Short> {
+public interface LocalLevelRepository extends JpaRepository<LocalLevel,Integer> {
 }
